fix: Conducteur poids datatype

+++ Conducteur.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public class Conducteur extends Voyageur{
     private int numeroDePermis;
 
-    public Conducteur(String prenom, String addresse, LocalDate dateDeNaissance, String nom, Double poids, LocalDate dateDeVoyage, int numeroDePermis) {
+    public Conducteur(String prenom, String addresse, LocalDate dateDeNaissance, String nom, int poids, LocalDate dateDeVoyage, int numeroDePermis) {
         super(prenom, addresse, dateDeNaissance, nom, poids, dateDeVoyage);
         this.numeroDePermis = numeroDePermis;
     }
